+++ core/app/models/spree/stock/estimator.rb
@@ -33,25 +33,13 @@ module Spree
       def calculate_shipping_rates(package)
         shipping_methods(package).map do |shipping_method|
           cost = shipping_method.calculator.compute(package)
-          tax_category = shipping_method.tax_category
-          if tax_category
-            tax_rate = tax_category.tax_rates.detect do |rate|
-              # If the rate's zone matches the order's zone, a positive adjustment will be applied.
-              # If the rate is from the default tax zone, then a negative adjustment will be applied.
-              # See the tests in shipping_rate_spec.rb for an example of this.d
-              rate.zone == package.shipment.order.tax_zone || rate.zone.default_tax?
-            end
-          end
-
           if cost
             rate = shipping_method.shipping_rates.new(
               cost: cost,
               shipment: package.shipment
             )
-            rate.tax_rate = tax_rate if tax_rate
+            Spree::Config.shipping_rate_taxer_class.new.tax(rate)
           end
-
-          rate
         end.compact
       end